+++ lib/engineyard-serverside/deploy.rb
@@ -382,6 +382,7 @@ YML
 
       def callback(what)
         @callbacks_reached ||= true
+
         if paths.deploy_hook(what).exist?
           shell.status "Running deploy hook: deploy/#{what}.rb"
           run Escape.shell_command(base_callback_command_for(what)) do |server, cmd|
@@ -390,6 +391,13 @@ YML
             per_instance_args << '--current-name'  << server.name.to_s if server.name
             per_instance_args << '--config'        << config.to_json
             cmd << " " << Escape.shell_command(per_instance_args)
+          end
+        elsif paths.executable_deploy_hook(what).executable?
+          shell.status "Running deploy hook: deploy/#{what}"
+          run [About.hook_executor, what.to_s].join(' ') do |server, cmd|
+            cmd = hook_env_vars(server).reject{|k,v| v.nil?}.map{|k,v|
+              "#{k}=#{Escape.shell_command([v])}"
+            }.join(' ') + ' ' + config.framework_envs + ' ' + cmd
           end
         end
       end
@@ -410,6 +418,20 @@ YML
         cmd << '--framework-env'    << config.framework_env.to_s
         cmd << '--verbose' if config.verbose
         cmd
+      end
+
+      def hook_env_vars(server)
+        {
+          'EY_DEPLOY_ACCOUNT_NAME' => config.account_name,
+          'EY_DEPLOY_APP' => config.app,
+          'EY_DEPLOY_CONFIG' => config.to_json,
+          'EY_DEPLOY_CURRENT_ROLES' => server.roles.to_a.join(' '),
+          'EY_DEPLOY_CURRENT_NAME' => server.name ? server.name.to_s : nil,
+          'EY_DEPLOY_ENVIRONMENT_NAME' => config.environment_name,
+          'EY_DEPLOY_FRAMEWORK_ENV' => config.framework_env.to_s,
+          'EY_DEPLOY_RELEASE_PATH' => paths.active_release.to_s,
+          'EY_DEPLOY_VERBOSE' => (config.verbose ? '1' : '0'),
+        }
       end
 
       # FIXME: Legacy method, warn and remove.